+++ app/assets/javascripts/fae/application.js
@@ -13,8 +13,8 @@
 //= require jquery
 //= require jquery_ujs
 //= require jquery.remotipart
-//= require jquery.ui.sortable
-//= require jquery.ui.datepicker
+//= require jquery-ui/widgets/sortable
+//= require jquery-ui/widgets/datepicker
 //= require underscore
 //= require json2
 //= require judge
@@ -24,6 +24,7 @@
 //= require fae/vendor/fryr
 //= require fae/vendor/moment.min
 //= require fae/vendor/jquery.daterangepicker
+//= require fae/vendor/jqColorPicker.min
 //= require fae/vendor/jquery.multi-select
 //= require fae/vendor/jquery.simplemodal.1.4.4
 //= require fae/vendor/jquery.tablesorter
@@ -36,6 +37,7 @@
 //= require fae/form/hinter
 //= require fae/form/_form
 //= require fae/form/inputs/_dates
+//= require fae/form/inputs/_color
 //= require fae/form/inputs/_select
 //= require fae/form/inputs/_text
 //= require fae/form/inputs/_checkbox